Block share create if there are no policies attached to consumption role
